+++ packages/cli/test/cli_test.js
@@ -11,6 +11,8 @@ const readConfigFile = configPath => {
   return require(configPath)
 }
 
+// Beware: These tests assume that the whole suite runs in order.
+// The tests build up on each other – keeping it pragmatic here.
 describe('CLI', function () {
   this.timeout(5000)
 
@@ -23,10 +25,7 @@ describe('CLI', function () {
 
       // stdout
       assertMatches(stdout, 'Initialized Cli Project')
-      assertMatches(stdout, 'The following files were created:')
-      assertMatches(stdout, 'uiengine.config.js')
-      assertMatches(stdout, 'src/uiengine/pages/page.md')
-      assertMatches(stdout, 'src/templates/uiengine.html')
+      assertMatches(stdout, 'The following files were created:\n\n- uiengine.config.js\n- src/templates/uiengine.html\n- src/uiengine/pages/page.md')
 
       // config
       const configPath = join(testPath, 'uiengine.config.js')
@@ -57,8 +56,28 @@ describe('CLI', function () {
       assertContentMatches(previewPath, 'add your custom scripts here')
     })
 
+    it('should not overwrite existing files', async () => {
+      const stdout = await runCommand(testPath, 'uiengine init')
+
+      // stdout
+      assertMatches(stdout, 'The following files already existed:\n\n- uiengine.config.js\n- src/templates/uiengine.html\n- src/uiengine/pages/page.md')
+    })
+
+    describe('with force flag', () => {
+      it('should overwrite existing files', async () => {
+        const stdout = await runCommand(testPath, 'uiengine init -f')
+
+        // stdout
+        assertMatches(stdout, 'The following files were created:\n\n- uiengine.config.js\n- src/templates/uiengine.html\n- src/uiengine/pages/page.md')
+      })
+    })
+
     describe('with override flag', () => {
       it('should override config parameters', async () => {
+        // remove files from previous run
+        removeSync(testTmpPath)
+        ensureDirSync(testPath)
+
         await runCommand(testPath, 'uiengine init --override.name=OVERRIDE --override.source.pages=uiengine/pages --override.target=./dist/override-target --override.ui.lang=de')
 
         const configPath = join(testPath, 'uiengine.config.js')
@@ -73,6 +92,10 @@ describe('CLI', function () {
 
     describe('with demo flag', () => {
       it('should create the demo files', async () => {
+        // remove files from previous run
+        removeSync(testTmpPath)
+        ensureDirSync(testPath)
+
         const stdout = await runCommand(testPath, 'uiengine init --demo')
 
         // stdout
@@ -100,28 +123,63 @@ describe('CLI', function () {
 
       // stdout
       assertMatches(stdout, 'Pages created')
-      assertMatches(stdout, 'The following files were created:')
-      assertMatches(stdout, atomsPagePath)
-      assertMatches(stdout, moleculesPagePath)
+      assertMatches(stdout, `The following files were created:\n\n- ${atomsPagePath}\n- ${moleculesPagePath}`)
 
       //  page files
       assertContentMatches(join(testPath, atomsPagePath), '# Atoms')
       assertContentMatches(join(testPath, moleculesPagePath), '# Molecules')
     })
+
+    it('should not overwrite existing files', async () => {
+      const stdout = await runCommand(testPath, 'uiengine page atoms molecules organisms')
+      const atomsPagePath = 'src/uiengine/pages/atoms/page.md'
+      const moleculesPagePath = 'src/uiengine/pages/molecules/page.md'
+      const organismsPagePath = 'src/uiengine/pages/organisms/page.md'
+
+      // stdout
+      assertMatches(stdout, `The following files already existed:\n\n- ${atomsPagePath}\n- ${moleculesPagePath}`)
+      assertMatches(stdout, `The following files were created:\n\n- ${organismsPagePath}`)
+    })
+
+    describe('with force flag', () => {
+      it('should overwrite existing files ', async () => {
+        const stdout = await runCommand(testPath, 'uiengine page atoms molecules --force')
+        const atomsPagePath = 'src/uiengine/pages/atoms/page.md'
+        const moleculesPagePath = 'src/uiengine/pages/molecules/page.md'
+
+        // stdout
+        assertMatches(stdout, `The following files were created:\n\n- ${atomsPagePath}\n- ${moleculesPagePath}`)
+      })
+    })
   })
 
   describe('component command', () => {
     it('should create the basic files for a new component', async () => {
+      const stdout = await runCommand(testPath, 'uiengine component list')
+
+      // stdout
+      assertMatches(stdout, 'List created')
+      assertMatches(stdout, `The following files were created:\n\n- src/components/list/component.md`)
+
+      // component file
+      const markdownPath = join(testPath, 'src/components/list/component.md')
+      assertContentMatches(markdownPath, 'title: List')
+    })
+
+    it('should not overwrite existing files', async () => {
       const stdout = await runCommand(testPath, 'uiengine component button default primary')
 
       // stdout
-      assertMatches(stdout, 'Button created')
-      assertMatches(stdout, 'The following files were created:')
-      assertMatches(stdout, 'src/components/button/component.md')
+      assertMatches(stdout, `The following files already existed:\n\n- src/components/button/component.md`)
+    })
 
-      // component file
-      const markdownPath = join(testPath, 'src/components/button/component.md')
-      assertContentMatches(markdownPath, 'title: Button')
+    describe('with force flag', () => {
+      it('should overwrite existing files ', async () => {
+        const stdout = await runCommand(testPath, 'uiengine component button --force')
+
+        // stdout
+        assertMatches(stdout, `The following files were created:\n\n- src/components/button/component.md`)
+      })
     })
   })