+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   - The html adapter is part of the standard install, but it is unlikely that you would want to generate the files for every new component.
 - Variant ID and filename contain the index
   - allows for rendering multiple variants based on the same file
+- CLI commands (init, component, page) do not overwrite existing files
+  - unless the `--force` flag is provided
 
 ### Fixed